Replaced PackageFileSerializer by implementations of JsonConverter
