+++ composer.json
@@ -14,7 +14,7 @@
         "puli/repository": "^1.0-beta9",
         "puli/discovery": "^1.0-beta9",
         "puli/url-generator": "^1.0-beta4",
-        "webmozart/json": "^1.2.1",
+        "webmozart/json": "^1.3@dev",
         "webmozart/path-util": "^2.2.3",
         "webmozart/expression": "^1.0",
         "webmozart/assert": "^1.0",
