--- src/Api/Migration/MigrationException.php
@@ -1,25 +1,0 @@
-<?php
-
-/*
- * This file is part of the puli/manager package.
- *
- * (c) Bernhard Schussek <bschussek@gmail.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Puli\Manager\Api\Migration;
-
-use RuntimeException;
-
-/**
- * Thrown when a migration fails.
- *
- * @since  1.0
- *
- * @author Bernhard Schussek <bschussek@gmail.com>
- */
-class MigrationException extends RuntimeException
-{
-}
